+++ include/server.h
@@ -58,8 +58,8 @@ struct ServerState {
       for (int i = 0; i < FISH_NUM; i++) {
           Fish f;
           f.spawn(
-             100 + rand() % (SCREEN_WIDTH - 200),
-              100 + rand() % (SCREEN_HEIGHT - 200) 
+            (float) 100 + rand() % (SCREEN_WIDTH - 200),
+            (float) 100 + rand() % (SCREEN_HEIGHT - 200) 
           );
           fishes[i] = f;
       }
